\name{affym}
\alias{affym}
\title{
Preprocessing and Analysis of Loaded Affymetrix Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Affymetrix data automatically if exists object of Affymetrix as datAffy.
}
\author{
Dhammapal Bharne
}
