\name{agtwo}
\alias{agtwo}
\title{
Preprocessing and Analysis of Loaded Agilent_TwoColor Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Agilent_TwoColor data automatically if exists object of Agilent OneColor as datAgTwo.
}
\author{
Dhammapal Bharne
}
