\name{clas}
\alias{clas}
\title{
Classification of Differentially Expression Genes
}
\description{
Performs Classification of Differentially Expression Genes
}
\details{
Function can perform classification of differentially expressed genes based on filtered data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .f data. It maps differentially expressed genes to corresponding normalized .m or .m2 data and classify the genes. The outputs are Clas_Affy for Affymetrix, Clas_Ag1 for Agilent_OneColor, Clas_Ag2 for Agilent_TwoColor, Clas_Il_B for Illumina_Beadarray, Clas_Il_L for Illumina_Lumi, Clas_N for Nimblegen, Clas_S for Series_Matrix and Clas_O for Online procedures.
}
\author{
Dhammapal Bharne
}
