\name{clust}
\alias{clust}
\title{
Clustering of Normalized Microarray Data
}
\description{
Performs Clustering of normalized microarray data
}
\details{
Function can perform clustering on normalized microarray data if exists object of Affymetrix data as dat2Affy.m, Agilent_OneColor as datAgOne2.m, Agilent_TwoColor as datAgTwo2.m, Illumina_Beadarray as datIllBA2.m2, Illumina_Lumi as lumi_NQ.m, Nimblegen as data.matrix_Nimblegen2.m, Series_Matrix as data.matrixNorm.m and Online_Data as data.matrix_onlineNorm.m. The outputs are sample.dist_Affy and sample.clust_Affy for Affymetrix, sample.dist_Ag1 and sample.clust_Ag1 for Agilent_OneColor, sample.dist_Ag2 and sample.clust_Ag2 for Agilent_TwoColor, sample.dist_Il_B and sample.clust_Il_B for Illumina_Beadarray, sample.dist_Il_L and sample.clust_Il_L for Illumina_Lumi, sample.dist_N and sample.clust_N for Nimblegen, sample.dist_S and sample.clust_S for Series_Matrix and sample.dist_O and sample.clust_O for Online procedures.
}
\author{
Dhammapal Bharne
}
