\name{dmo}
\alias{dmo}
\title{
Preprocessing and Analysis of Loaded Online Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Online data automatically if exists object of Online_Data as data.matrix_onlineImp.
}
\author{
Dhammapal Bharne
}
