\name{dmsm}
\alias{dmsm}
\title{
Preprocessing and Analysis of Loaded Series_Matrix Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Series_Matrix data automatically if exists object data.matrixImp of Series_Matrix data.
}
\author{
Dhammapal Bharne
}
