\name{graph_gsea_goBP}
\alias{graph_gsea_goBP}
\title{
Plots Graph of GO Terms for Biological Process of GSEA
}
\description{
Plots a network graph of GO terms for Biological Process obtained from gene set enrichment analysis.
}
\details{
Function can plot a graph for GO terms obtained from gene set enrichment analysis. It is a network of GO terms of Biological Process as nodes and links as directed lines. It is performed if exists object GOresultBP_Affy for Affymetrix, GOresultBP_Ag1 for Agilent_OneColor, GOresultBP_Ag2 for Agilent_TwoColor, GOresultBP_Il_B for Illumina_Beadarray, GOresultBP_Il_L for Illumina_Lumi, GOresultBP_N for Nimblegen, GOresultBP_S for Series_Matrix and GOresultBP_O for Online procedures.
}
\author{
Dhammapal Bharne
}
