\name{graph_gsea_goCC}
\alias{graph_gsea_goCC}
\title{
Plots Graph of GO Terms for Cellular Component of GSEA
}
\description{
Plots a network graph of GO terms for Cellular Component obtained from gene set enrichment analysis.
}
\details{
Function can plot a graph for GO terms obtained from gene set enrichment analysis. It is a network of GO terms of Cellular Component as nodes and links as directed lines. It is performed if exists object GOresultCC_Affy for Affymetrix, GOresultCC_Ag1 for Agilent_OneColor, GOresultCC_Ag2 for Agilent_TwoColor, GOresultCC_Il_B for Illumina_Beadarray, GOresultCC_Il_L for Illumina_Lumi, GOresultCC_N for Nimblegen, GOresultCC_S for Series_Matrix and GOresultCC_O for Online procedures.
}
\author{
Dhammapal Bharne
}
