\name{graph_gsea_goMF}
\alias{graph_gsea_goMF}
\title{
Plots Graph of GO Terms for Molecular Function of GSEA
}
\description{
Plots a network graph of GO terms for Molecular Function obtained from gene set enrichment analysis.
}
\details{
Function can plot a graph for GO terms obtained from gene set enrichment analysis. It is a network of GO terms of Molecular Function as nodes and links as directed lines. It is performed if exists object GOresultMF_Affy for Affymetrix, GOresultMF_Ag1 for Agilent_OneColor, GOresultMF_Ag2 for Agilent_TwoColor, GOresultMF_Il_B for Illumina_Beadarray, GOresultMF_Il_L for Illumina_Lumi, GOresultMF_N for Nimblegen, GOresultMF_S for Series_Matrix and GOresultMF_O for Online procedures.
}
\author{
Dhammapal Bharne
}
