\name{graph_gsea_kegg}
\alias{graph_gsea_kegg}
\title{
Plots Graph of Pathways for a KEGG ID from GSEA
}
\description{
Plots a network graph of KEGG pathways for a KEGG ID obtained from gene set enrichment analysis.
}
\details{
Function can plot a graph of KEGG Pathways for KEGG ID obtained from gene set enrichment analysis. It is a network of genes as nodes and links as directed lines. It is performed if exists object KEGGresult_Affy for Affymetrix, KEGGresult_Ag1 for Agilent_OneColor, KEGGresult_Ag2 for Agilent_TwoColor, KEGGresult_Il_B for Illumina_Beadarray, KEGGresult_Il_L for Illumina_Lumi, KEGGresult_N for Nimblegen, KEGGresult_S for Series_Matrix and KEGGresult_O for Online procedures.
}
\author{
Dhammapal Bharne
}
