\name{graph_gsta_goBP}
\alias{graph_gsta_goBP}
\title{
Plots Graph of GO Terms for Biological Process of GSTA
}
\description{
Plots a network graph of GO terms for Biological Process obtained from gene set test analysis.
}
\details{
Function can plot a graph for GO terms obtained from gene set test analysis. It is a network of GO terms of Biological Process as nodes and links as directed lines. It is performed if exists object GOtable.outBP_Affy for Affymetrix, GOtable.outBP_Ag1 for Agilent_OneColor, GOtable.outBP_Ag2 for Agilent_TwoColor, GOtable.outBP_Il_B for Illumina_Beadarray, GOtable.outBP_Il_L for Illumina_Lumi, GOtable.outBP_N for Nimblegen, GOtable.outBP_S for Series_Matrix and GOtable.outBP_O for Online procedures.
}
\author{
Dhammapal Bharne
}
