\name{graph_gsta_goCC}
\alias{graph_gsta_goCC}
\title{
Plots Graph of GO Terms for Cellular Component of GSTA
}
\description{
Plots a network graph of GO terms for Cellular Component obtained from gene set test analysis.
}
\details{
Function can plot a graph for GO terms obtained from gene set test analysis. It is a network of GO terms of Cellular Component as nodes and links as directed lines. It is performed if exists object GOtable.outCC_Affy for Affymetrix, GOtable.outCC_Ag1 for Agilent_OneColor, GOtable.outCC_Ag2 for Agilent_TwoColor, GOtable.outCC_Il_B for Illumina_Beadarray, GOtable.outCC_Il_L for Illumina_Lumi, GOtable.outCC_N for Nimblegen, GOtable.outCC_S for Series_Matrix and GOtable.outCC_O for Online procedures.
}
\author{
Dhammapal Bharne
}
