\name{graph_gsta_goMF}
\alias{graph_gsta_goMF}
\title{
Plots Graph of GO Terms for Molecular Function of GSTA
}
\description{
Plots a network graph of GO terms for Molecular Function obtained from gene set test analysis.
}
\details{
Function can plot a graph for GO terms obtained from gene set test analysis. It is a network of GO terms of Molecular Function as nodes and links as directed lines. It is performed if exists object GOtable.outMF_Affy for Affymetrix, GOtable.outMF_Ag1 for Agilent_OneColor, GOtable.outMF_Ag2 for Agilent_TwoColor, GOtable.outMF_Il_B for Illumina_Beadarray, GOtable.outMF_Il_L for Illumina_Lumi, GOtable.outMF_N for Nimblegen, GOtable.outMF_S for Series_Matrix and GOtable.outMF_O for Online procedures.
}
\author{
Dhammapal Bharne
}
