\name{graph_gsta_kegg}
\alias{graph_gsta_kegg}
\title{
Plots Graph of Pathways for a KEGG ID from GSTA
}
\description{
Plots a network graph of KEGG pathways for a KEGG ID obtained from gene set test analysis.
}
\details{
Function can plot a graph of KEGG Pathways for KEGG ID obtained from gene set test analysis. It is a network of genes as nodes and links as directed lines. It is performed if exists object KEGGtable.out_Affy for Affymetrix, KEGGtable.out_Ag1 for Agilent_OneColor, KEGGtable.out_Ag2 for Agilent_TwoColor, KEGGtable.out_Il_B for Illumina_Beadarray, KEGGtable.out_Il_L for Illumina_Lumi, KEGGtable.out_N for Nimblegen, KEGGtable.out_S for Series_Matrix and KEGGtable.out_O for Online procedures.
}
\author{
Dhammapal Bharne
}
