\name{gsea}
\alias{gsea}
\title{
Gene Set Enrichment Analysis of Microarray Data
}
\description{
Performs Gene Set Enrichment Analysis of Microarray Data. It include GO categories and KEGG pathways.
}
\details{
Function can perform gene set enrichment analysis of statistical significant microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .s data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are html reports.
}
\author{
Dhammapal Bharne
}
