\name{gsea_goBP}
\alias{gsea_goBP}
\title{
Gene Set Enrichment Analysis for Biological Process
}
\description{
Performs Gene Set Enrichment Analysis on statistically significant microarray data to obtain GO terms of Biological Process.
}
\details{
Function can perform gene set enrichment analysis of statistical significant microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .s data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are GOresultBP_Affy for Affymetrix, GOresultBP_Ag1 for Agilent_OneColor, GOresultBP_Ag2 for Agilent_TwoColor, GOresultBP_Il_B for Illumina_Beadarray, GOresultBP_Il_L for Illumina_Lumi, GOresultBP_N for Nimblegen, GOresultBP_S for Series_Matrix and GOresultBP_O for Online procedures.
}
\author{
Dhammapal Bharne
}
