\name{gsea_goCC}
\alias{gsea_goCC}
\title{
Gene Set Enrichment Analysis for Cellular Component
}
\description{
Performs Gene Set Enrichment Analysis on statistically significant microarray data to obtain GO terms of Cellular Component.
}
\details{
Function can perform gene set enrichment analysis of statistical significant microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .s data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are GOresultCC_Affy for Affymetrix, GOresultCC_Ag1 for Agilent_OneColor, GOresultCC_Ag2 for Agilent_TwoColor, GOresultCC_Il_B for Illumina_Beadarray, GOresultCC_Il_L for Illumina_Lumi, GOresultCC_N for Nimblegen, GOresultCC_S for Series_Matrix and GOresultCC_O for Online procedures.
}
\author{
Dhammapal Bharne
}
