\name{gsea_goMF}
\alias{gsea_goMF}
\title{
Gene Set Enrichment Analysis for Molecular Function
}
\description{
Performs Gene Set Enrichment Analysis on statistically significant microarray data to obtain GO terms of Molecular Function.
}
\details{
Function can perform gene set enrichment analysis of statistical significant microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .s data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are GOresultMF_Affy for Affymetrix, GOresultMF_Ag1 for Agilent_OneColor, GOresultMF_Ag2 for Agilent_TwoColor, GOresultMF_Il_B for Illumina_Beadarray, GOresultMF_Il_L for Illumina_Lumi, GOresultMF_N for Nimblegen, GOresultMF_S for Series_Matrix and GOresultMF_O for Online procedures.
}
\author{
Dhammapal Bharne
}
