\name{gsea_kegg}
\alias{gsea_kegg}
\title{
Gene Set Enrichment Analysis for KEGG Pathways
}
\description{
Performs Gene Set Enrichment Analysis on statistically significant microarray data to obtain KEGG pathways.
}
\details{
Function can perform gene set enrichment analysis of statistical significant microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .s data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are KEGGresult_Affy for Affymetrix, KEGGresult_Ag1 for Agilent_OneColor, KEGGresult_Ag2 for Agilent_TwoColor, KEGGresult_Il_B for Illumina_Beadarray, KEGGresult_Il_L for Illumina_Lumi, KEGGresult_N for Nimblegen, KEGGresult_S for Series_Matrix and KEGGresult_O for Online procedures.
}
\author{
Dhammapal Bharne
}
