\name{gsta}
\alias{gsta}
\title{
Gene Set Test Analysis of Microarray Data
}
\description{
Performs Gene Set Test Analysis of Microarray Data. It include GO categories and KEGG pathways.
}
\details{
Function can perform gene set test analysis of normalized microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .m or .m2 data. It needs GEOmetadb.sqlite database file or internet connection.
}
\author{
Dhammapal Bharne
}
