\name{gsta_goBP}
\alias{gsta_goBP}
\title{
Gene Set Test Analysis for Biological Process
}
\description{
Performs Gene Set Test Analysis on normalized microarray data to obtain GO terms of Biological Process.
}
\details{
Function can perform gene set test analysis of normalized microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .m or m2 data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are GOtable.outBP_Affy for Affymetrix, GOtable.outBP_Ag1 for Agilent_OneColor, GOtable.outBP_Ag2 for Agilent_TwoColor, GOtable.outBP_Il_B for Illumina_Beadarray, GOtable.outBP_Il_L for Illumina_Lumi, GOtable.outBP_N for Nimblegen, GOtable.outBP_S for Series_Matrix and GOtable.outBP_O for Online procedures.
}
\author{
Dhammapal Bharne
}
