\name{gsta_goCC}
\alias{gsta_goCC}
\title{
Gene Set Test Analysis for Cellular Component
}
\description{
Performs Gene Set Test Analysis on normalized microarray data to obtain GO terms of Cellular Component.
}
\details{
Function can perform gene set test analysis of normalized microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .m or m2 data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are GOtable.outCC_Affy for Affymetrix, GOtable.outCC_Ag1 for Agilent_OneColor, GOtable.outCC_Ag2 for Agilent_TwoColor, GOtable.outCC_Il_B for Illumina_Beadarray, GOtable.outCC_Il_L for Illumina_Lumi, GOtable.outCC_N for Nimblegen, GOtable.outCC_S for Series_Matrix and GOtable.outCC_O for Online procedures.
}
\author{
Dhammapal Bharne
}
