\name{gsta_goMF}
\alias{gsta_goMF}
\title{
Gene Set Test Analysis for Molecular Function
}
\description{
Performs Gene Set Test Analysis on normalized microarray data to obtain GO terms of Molecular Function.
}
\details{
Function can perform gene set test analysis of normalized microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .m or m2 data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are GOtable.outMF_Affy for Affymetrix, GOtable.outMF_Ag1 for Agilent_OneColor, GOtable.outMF_Ag2 for Agilent_TwoColor, GOtable.outMF_Il_B for Illumina_Beadarray, GOtable.outMF_Il_L for Illumina_Lumi, GOtable.outMF_N for Nimblegen, GOtable.outMF_S for Series_Matrix and GOtable.outMF_O for Online procedures.
}
\author{
Dhammapal Bharne
}
