\name{gsta_kegg}
\alias{gsta_kegg}
\title{
Gene Set Test Analysis for KEGG Pathways
}
\description{
Performs Gene Set Test Analysis on normalized microarray data to obtain KEGG pathways.
}
\details{
Function can perform gene set test analysis of normalized microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .m or m2 data. It needs GEOmetadb.sqlite database file or internet connection. The outputs are KEGGtable.out_Affy for Affymetrix, KEGGtable.out_Ag1 for Agilent_OneColor, KEGGtable.out_Ag2 for Agilent_TwoColor, KEGGtable.out_Il_B for Illumina_Beadarray, KEGGtable.out_Il_L for Illumina_Lumi, KEGGtable.out_N for Nimblegen, KEGGtable.out_S for Series_Matrix and KEGGtable.out_O for Online procedures.
}
\author{
Dhammapal Bharne
}
