\name{illbead}
\alias{illbead}
\title{
Preprocessing and Analysis of Loaded Illumina_Beadarray Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Illumina_Beadarray data automatically if exists object of Illumina_Beadarray as datIllBA.
}
\author{
Dhammapal Bharne
}
