\name{illlumi}
\alias{illlumi}
\title{
Preprocessing and Analysis of Loaded Illumina_Lumi Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Illumina_Lumi data automatically if exists object of Illumina_Lumi as lumi_data.
}
\author{
Dhammapal Bharne
}
