\name{maGUI-package}
\alias{maGUI-package}
\docType{package}
\title{
A Graphical User Interface for Microarray Data Analysis and Annotation
}
\description{
Provides a comprehensive graphical user interface for analysis of Affymetrix, Agilent, Illumina, Nimblegen and other microarray data. It can perform miscellaneous tasks such gene set enrichment and test analysis, identifying gene symbols and building co-expression network. It can also estimate sample size for atleast two-fold expression change. Further, it can identify protein-protein associations from two normalized microarray data. The current version is its slenderized form for easy launching and implementation.
}
\details{
\tabular{ll}{
Package: \tab maGUI\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 2020-10-27\cr
License: \tab GPL-2\cr
}
}
\author{
Dhammapal Bharne, Vaibhav Vindal

Maintainer: Dhammapal Bharne <dhammapalb@gmail.com>
}
