\name{nimblg}
\alias{nimblg}
\title{
Preprocessing and Analysis of Loaded Nimblegen Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Nimblegen data automatically if exists object of Nimblegen as data.matrix_Nimblegen.
}
\author{
Dhammapal Bharne
}
