\name{ntwk}
\alias{ntwk}
\title{
Co-regulated Network for Differentially Expressed Genes
}
\description{
Builds Co-regulated Network for Differentially Expressed Genes.
}
\details{
Function can build co-regulated network for differentially expressed genes if exists object of Affymetrix data as DE_Affy, Agilent_OneColor as DE_Ag1, Agilent_TwoColor as DE_Ag2, Illumina_Beadarray as DE_Il_B, Illumina_Lumi as DE_Il_L, Nimblegen as DE_N, Series_Matrix as DE_S and Online_Data as DE_O. It maps differentially expressed genes to corresponding normalized .m or .m2 data, create adjacency matrix and builds the co-regulated newtwork.
}
\author{
Dhammapal Bharne
}
