\name{pca}
\alias{pca}
\title{
Pricipal Component Analysis on Normalized Microarray Data
}
\description{
Performs Principal Component Analysis on normalized microarray data
}
\details{
Function can perform pricipal component analysis on normalized microarray data if exists object of Affymetrix data as dat2Affy.m, Agilent_OneColor as datAgOne2.m, Agilent_TwoColor as datAgTwo2.m, Illumina_Beadarray as datIllBA2.m2, Illumina_Lumi as lumi_NQ.m, Nimblegen as data.matrix_Nimblegen2.m, Series_Matrix as data.matrixNorm.m and Online_Data as data.matrix_onlineNorm.m. The outputs are pca_Affy for Affymetrix, pca_Ag1 for Agilent_OneColor, pca_Ag2 for Agilent_TwoColor, pca_Il_B for Illumina_Beadarray, pca_Il_L for Illumina_Lumi, pca_N for Nimblegen, pca_S for Series_Matrix and pca_O for Online procedures.
}
\author{
Dhammapal Bharne
}
