\name{ppis}
\alias{ppis}
\title{
Protein-Protein Interactions Prediction from Microarray Data
}
\description{
Predicts-protein protein interactions from two normalized microarray data

}
\details{
Function can predict protein-protein interactions from two normalized microarray data if exists Affymetrix data as dat2Affy.m, Agilent_OneColor as datAgOne2.m, Agilent_TwoColor as datAgTwo2.m, Illumina_Beadarray as datIllBA2.m2, Illumina_Lumi as lumi_NQ.m, Nimblegen as data.matrix_Nimblegen2.m, Series_Matrix as data.matrixNorm.m and Online_Data as data.matrix_onlineNorm.m. The outputs are protein-protein pairs written to a file.
}
\author{
Dhammapal Bharne
}
