\name{sse}
\alias{sse}
\title{
Sample Size Estimation on Normalized Microarray Data
}
\description{
Performs Sample Size Estimation on Normalized Microarray Data
}
\details{
Function can perform estimation of sample size for atleast two-fold expression change, on normalized microarray data if exists object of Affymetrix data as dat2Affy.m, Agilent_OneColor as datAgOne2.m, Agilent_TwoColor as datAgTwo2.m, Illumina_Beadarray as datIllBA2.m2, Illumina_Lumi as lumi_NQ.m, Nimblegen as data.matrix_Nimblegen2.m, Series_Matrix as data.matrixNorm.m and Online_Data as data.matrix_onlineNorm.m. The outputs are size_Affy for Affymetrix, size_Ag1 for Agilent_OneColor, size_Ag2 for Agilent_TwoColor, size_Il_B for Illumina_Beadarray, size_Il_L for Illumina_Lumi, size_N for Nimblegen, size_S for Series_Matrix and size_O for Online procedures.
}
\author{
Dhammapal Bharne
}
