\name{stat_sign}
\alias{stat_sign}
\title{
Statisitical Significance of Filtering Data
}
\description{
Performs Statisitical Significance of Filtering Data.
}
\details{
Function can perform statistical significance test of filtered data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .f data. It maps toptable genes with p-value cut of 0.01 to corresponding normalized .m or .m2 data. The output data objects are the substitution of .m or .m2 to .s.
}
\author{
Dhammapal Bharne
}
