\name{symbol}
\alias{symbol}
\title{
Mapping of Identifiers of Microarray Data to Genes
}
\description{
Maps Identifiers of Microarray Data to Genes
}
\details{
Function can perform mapping of genes for identifiers of normalized microarray data if exists object of Affymetrix, Agilent_OneColor, Agilent_TwoColor, Illumina_Beadarray, Illumina_Lumi, Nimblegen, Series_Matrix and Online as .m or .m2 data. It needs GEOmetadb.sqlite database file or internet connection.
}
\author{
Dhammapal Bharne
}
