\name{agone}
\alias{agone}
\title{
Preprocessing and Analysis of Loaded Agilent_OneColor Data
}
\description{
Performs pre-processing and analysis of loaded microarray data automatically. It includes normalization, quality check, filtering, statistical significance, differential gene expression, principal component analysis, clustering and classification.
}
\details{
Function performs pre-processing and analysis of loaded Agilent_OneColor data automatically if exists object datAgOne of Agilent OneColor data.
}
\author{
Dhammapal Bharne
}
