\name{maGUI-package}
\alias{maGUI-package}
\docType{package}
\title{
A Graphical User Interface for Microarray Data Analysis and Annotation
}
\description{
Provides a graphical user interface for pre-processing and analyses of DNA microarray data of various platforms. It performs functional annotation and pathway analysis of differentially expressed genes. It identifies gene symbols, performs sample size estimation and also builds co-expression network.
}
\details{
\tabular{ll}{
Package: \tab maGUI\cr
Type: \tab Package\cr
Version: \tab 4.0\cr
Date: \tab 2022-04-12\cr
License: \tab GPL-2\cr
}
}
\author{
Dhammapal Bharne

Maintainer: Dhammapal Bharne <dhammapalb@gmail.com>
}
