\name{maGUI-package}
\alias{maGUI-package}
\docType{package}
\title{
A Graphical User Interface for Microarray Data Analysis
}
\description{
Provides a comprehensive graphical user interface for analysis of Affymetrix, Agilent, Illumina, Nimblegen and other microarray data. It can perform miscellaneous tasks such as gene set enrichment and test analysis, annotations and identifying gene symbols. It can also estimate sample size for atleast two-fold expression change. Further, it can predict protein-protein interactions from two normalized microarray data.  
}
\details{
\tabular{ll}{
Package: \tab maGUI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-24\cr
License: \tab GPL-2\cr
}
}
\author{
Dhammapal Bharne, Praveen Kant, Vaibhav Vindal

Maintainer: Dhammapal Bharne <dhammapalb@uohyd.ac.in>
}
